Update part6b.md
I suggest you add to Exercise 6.12 what a notification message should contain as this is not clear to the reader. Is it just the new anecdote text for a new anecdote, and the anecdote text and the voted for a vote action? Examples of how a notification message should look like would also be helpful.